requirements.txt
certifi
colorama
cycler
joblib
kiwisolver
matplotlib
numpy
olefile
pandas
Pillow
pyparsing
python-dateutil
pytz
scikit-learn
scipy
six
scikit-learn
threadpoolctl
torchaudio
torchvision
tqdm
typing-extensions
wincertstore
xlsxwriter
openpyxl
